+++ tests/testthat/test-gt_info.R
@@ -5,7 +5,7 @@ test_that('Test gt info', {
 
   gt_info_test <- iris |> df_info() |> gt_info(df_name = 'iris')
   expect_equal(gt_info_test |> class(), c('gt_tbl', 'list'))
-  expect_equal(gt_info_test |> length(), 18)
+  expect_equal(gt_info_test |> length(), 21)
 
   expect_equal(gt_info_test$`_data`$var, names(iris))
   expect_equal(gt_info_test$`_data`$type, sapply(iris, typeof) |> unname())